Allow kinetic libraries to read duplicate entries
By default, RMG forbids this. A reaction library is only allowed to have duplicates if they're all Arrhenius, and then it combines them into a MultiArrhenius. But sometimes, you just want to read an experimental/in-development library as written with the duplicates. This adds a relatively hidden developer/debug option of turning off the no duplicates requirement.